requirements.txt
pydantic==2.4.0
PyYAML==6.0
ruff==0.11.11
pytest==8.3.5